Reference

Base.close — Method

Base.close(x::JArray)

Jets.getblock — Method

Jets.getblock(x::JArray{T,N,A}, iblock::CartesianIndex) where {T,N,A}

Get a block from a remote worked based on the iblock index

Jets.setblock! — Method

setblock_from_id!(x::JArray, xblock, iblock::CartesianIndex)

Set a block from remote worker based on the blockmap id, and block index

JournaledJets.block_delete_fromid! — Method

block_delete_fromid!(_id, _whence, iblock)

Remove a block from the registry based on the block map id and block key

JournaledJets.blocklength_from_id — Method

blocklength_from_id(id, iblock)

compute local block length based on blockmap id and local block id

JournaledJets.blocksize_from_id — Method

blocksize_from_id(id, iblock, idim)

compute local block size based on blockmap id and local block id and a given dimension

JournaledJets.close_by_id — Method

close_by_id(id)

Delete the work id from the registry if it hasnt already been deleted. Used when work finishes

JournaledJets.fill_local_part — Method

initialize_local_part(id, pids, nblocks::NTuple{N,Int}, ::Type{A}) where {N,A}

Create an empty block map for a given id in the registry

JournaledJets.finish_local_part — Method

finish_local_part(id, indices)

JournaledJets.getblock_and_delete_fromid! — Method

getblock_and_delete_fromid!(_id, _whence, iblock)

Pop a block off the registry based on the block map id and block key

JournaledJets.getblockmap_from_id! — Method

getblockmap_from_id!(id, _blockmap)

Pull out a block map from the registry based on id

JournaledJets.getindex_from_id — Method

getindex_from_id(id, iblock, i, N)

JournaledJets.getindices — Method

getindices(id, nblocks::NTuple{1}, blockmap)

JournaledJets.getindices — Method

getindices(id, nblocks::NTuple{N}, blockmap) where {N}

JournaledJets.indices — Method

indices(A::JArray)

return indices from JArray passed in

JournaledJets.indices_from_id! — Method

indices_from_id!(id, _indices)

return indices from registry based on passed in id

JournaledJets.initialize_local_part — Method

initialize_local_part(id, pids, nblocks::NTuple{N,Int}, ::Type{A}) where {N,A}

Create an empty block map for a given id in the registry

JournaledJets.setblock_from_id! — Method

setblock_from_id!(id, xblock, iblock)

Set a block in the registry based on the blockmap id, and block index

JournaledJets.setblockmap_from_id! — Method

setblockmap_from_id!(id, _blockmap)

Add a block map to the registry based using id as the 'key'

JournaledJets.update_blockmap_and_pids_fromid! — Method

update_blockmap_and_pids_fromid!(_id, _whence, iblock)

Update a block in the registry based on the block map id and block key with _whence

Serialization.deserialize — Method

Serialization.deserialize(S::AbstractSerializer, ::Type{J}) where {T,N,A,J<:JArray{T,N,A}}

Deserilization implementation for JArray struct

Serialization.serialize — Method

Serialization.serialize(S::AbstractSerializer, x::JArray{T,N}) where {T,N}

Serilization implementation for JArray struct

Index

Base.close

Jets.getblock

Jets.setblock!

JournaledJets.block_delete_fromid!

JournaledJets.blocklength_from_id

JournaledJets.blocksize_from_id

JournaledJets.close_by_id

JournaledJets.fill_local_part

JournaledJets.finish_local_part

JournaledJets.getblock_and_delete_fromid!

JournaledJets.getblockmap_from_id!

JournaledJets.getindex_from_id

JournaledJets.getindices

JournaledJets.getindices

JournaledJets.indices

JournaledJets.indices_from_id!

JournaledJets.initialize_local_part

JournaledJets.setblock_from_id!

JournaledJets.setblockmap_from_id!

JournaledJets.update_blockmap_and_pids_fromid!

Serialization.deserialize

Serialization.serialize